requirements.txt
Flask-SQLAlchemy
Flask-JWT
pytest
requests
